core: rename page component folders
