+++ Blasphemous.Modding.Installer/PageComponents/Loaders/ILoader.cs
@@ -1,0 +1,6 @@
+﻿namespace Blasphemous.Modding.Installer.PageComponents.Loaders;
+
+internal interface ILoader
+{
+    void LoadAllData();
+}
