+++ Blasphemous.Modding.Installer/PageComponents/Loaders/SkinLoader.cs
@@ -1,9 +1,9 @@
-﻿using Blasphemous.Modding.Installer.PageComponents.Sorting;
-using Blasphemous.Modding.Installer.PageComponents.UIHolding;
+﻿using Blasphemous.Modding.Installer.PageComponents.Sorters;
+using Blasphemous.Modding.Installer.PageComponents.UIHolders;
 using Blasphemous.Modding.Installer.Skins;
 using Newtonsoft.Json;
 
-namespace Blasphemous.Modding.Installer.PageComponents.Loading;
+namespace Blasphemous.Modding.Installer.PageComponents.Loaders;
 
 internal class SkinLoader : ILoader
 {
