--- Blasphemous.Modding.Installer/PageComponents/Loading/ILoader.cs
@@ -1,6 +1,0 @@
-﻿namespace Blasphemous.Modding.Installer.PageComponents.Loading;
-
-internal interface ILoader
-{
-    void LoadAllData();
-}
