+++ Blasphemous.Modding.Installer/PageComponents/Sorters/ISorter.cs
@@ -1,0 +1,6 @@
+﻿namespace Blasphemous.Modding.Installer.PageComponents.Sorters;
+
+internal interface ISorter
+{
+    void Sort();
+}
