--- Blasphemous.Modding.Installer/PageComponents/Sorting/ISorter.cs
@@ -1,6 +1,0 @@
-﻿namespace Blasphemous.Modding.Installer.PageComponents.Sorting;
-
-internal interface ISorter
-{
-    void Sort();
-}
