+++ Blasphemous.Modding.Installer/PageComponents/UIHolders/GenericUIHolder.cs
@@ -1,4 +1,4 @@
-﻿namespace Blasphemous.Modding.Installer.PageComponents.UIHolding;
+﻿namespace Blasphemous.Modding.Installer.PageComponents.UIHolders;
 
 internal class GenericUIHolder<T> : IUIHolder
 {
